+++ pkg/bpmn_engine/task_handler.go
@@ -7,8 +7,10 @@ type taskMatcher func(element *BPMN20.TaskElement) bool
 type taskHandlerType string
 
 const (
-	taskHandlerForId   = "TASK_HANDLER_ID"
-	taskHandlerForType = "TASK_HANDLER_TYPE"
+	taskHandlerForId              = "TASK_HANDLER_ID"
+	taskHandlerForType            = "TASK_HANDLER_TYPE"
+	taskHandlerForAssignee        = "TASK_HANDLER_ASSIGNEE"
+	taskHandlerForCandidateGroups = "TASK_HANDLER_CANDIDATE_GROUPS"
 )
 
 type taskHandler struct {
@@ -37,6 +39,15 @@ type NewTaskHandlerCommand1 interface {
 	// Hereby 'type' is defined as 'taskDefinition' extension element in the BPMN file.
 	// This allows a single handler to be used for multiple task definitions.
 	Type(taskType string) NewTaskHandlerCommand2
+
+	// Assignee defines a handler for a User Task with a given 'assignee';
+	// Hereby 'assignee' is defined as 'assignmentDefinition' extension element in the BPMN file.
+	Assignee(assignee string) NewTaskHandlerCommand2
+
+	// CandidateGroups defines a handler for a User Task with given 'candidate groups';
+	// For the handler you can specify one or more groups.
+	// If at least one matches a given user task, the handler will be called.
+	CandidateGroups(groups ...string) NewTaskHandlerCommand2
 }
 
 // NewTaskHandler registers a handler function to be called for service tasks with a given taskId
@@ -77,18 +88,43 @@ func (thc newTaskHandlerCommand) Handler(f func(job ActivatedJob)) {
 	thc.append(&th)
 }
 
-func (state *BpmnEngineState) findTaskHandler(element *BPMN20.TaskElement) func(job ActivatedJob) {
-	for _, handler := range state.taskHandlers {
-		if handler.handlerType == taskHandlerForId {
-			if handler.matches(element) {
-				return handler.handler
+// Assignee implements NewTaskHandlerCommand2
+func (thc newTaskHandlerCommand) Assignee(assignee string) NewTaskHandlerCommand2 {
+	thc.matcher = func(element *BPMN20.TaskElement) bool {
+		return (*element).GetAssignmentAssignee() == assignee
+	}
+	thc.handlerType = taskHandlerForAssignee
+	return thc
+}
+
+// CandidateGroups implements NewTaskHandlerCommand2
+func (thc newTaskHandlerCommand) CandidateGroups(groups ...string) NewTaskHandlerCommand2 {
+	thc.matcher = func(element *BPMN20.TaskElement) bool {
+		for _, group := range groups {
+			if contains((*element).GetAssignmentCandidateGroups(), group) {
+				return true
 			}
 		}
+		return false
 	}
-	for _, handler := range state.taskHandlers {
-		if handler.handlerType == taskHandlerForType {
-			if handler.matches(element) {
-				return handler.handler
+	thc.handlerType = taskHandlerForCandidateGroups
+	return thc
+}
+
+func (state *BpmnEngineState) findTaskHandler(element *BPMN20.TaskElement) func(job ActivatedJob) {
+	searchOrder := []taskHandlerType{taskHandlerForId}
+	if (*element).GetType() == BPMN20.ServiceTask {
+		searchOrder = append(searchOrder, taskHandlerForType)
+	}
+	if (*element).GetType() == BPMN20.UserTask {
+		searchOrder = append(searchOrder, taskHandlerForAssignee, taskHandlerForCandidateGroups)
+	}
+	for _, handlerType := range searchOrder {
+		for _, handler := range state.taskHandlers {
+			if handler.handlerType == handlerType {
+				if handler.matches(element) {
+					return handler.handler
+				}
 			}
 		}
 	}