+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## v0.?.?
 
 - support handlers being registered for (task definition) types (#58 BREAKING CHANGE)
+- support handlers for user tasks being registered for assignee or candidate groups (#59)
 
 ## Migration notes for breaking changes